refactor: extract PlaylistRow and add test
